Split preset into 2 files
